Refactor routes handler into focused modules (subtask 9.1)
Split src/api/handlers/routes.rs (1187 lines) into modular structure:
- routes/mod.rs: Handler functions and tests (546 lines)
- routes/types.rs: DTOs and XDS conversions (454 lines)
- routes/validation.rs: Validation utilities (235 lines)

Benefits:
- All modules under 500-line target for maintainability
- Clear separation of concerns (handlers/types/validation)
- Improved code organization and readability
- All 299 tests passing
- No breaking changes to public API
